Better Stack Traces ===================
Invariants are now using the stack captured when Assert was called,
instead of creating a new Error when evaulated. This should better
pinpoint the file:row:col where the error is actually at.

We're now partial way to lazy-running our invariants. Since I'm not
overwriting the predicate any longer, I was able to illiminate that part
of the code, and incorporate it into the invariants run() method.

Changes to be committed:
------------------------

- Modified:   src/assert/assert.js
- Modified:   src/invariant/invariant.js
- Modified:   src/invariant/invariant.test.js
- Deleted:    src/invariant/predicate.js
- Deleted:    src/invariant/predicate.test.js
- Modified:   src/plan/plan.test.js
- Modified:   src/reporter/reporter.js