Invariants are Lazy-Evaluated =============================
Ivariants are no longer evaluated immediately when they are created.
Rather, they have a run() method that will invoke the predicate (with
the proper arguments, etc.) at a later time. This is a prerequisite to
adding async tests, promise support in actual/expected values, etc.

This also allowed me to skip evaluation entirely (the predicate function
is never run), and to flip the results from Success to Fail, and
vis-versa on the Invariant itself. Before this logic was part of
the Reporter.

To evaluate the Invariants, I've added an execute method to plans.
Currently, I'm modifying the Plan in-place. This lets me use current
logic as it exists in Reporter, etc., but I may decide to return an
array from execute, or append a results array to the plan. I'm
undecided.

Changes to be committed:
------------------------

- Modified:   src/assert/assert.js
- Modified:   src/assert/assert.test.js
- Modified:   src/invariant/index.js
- Modified:   src/invariant/invariant.js
- Modified:   src/invariant/invariant.test.js
- Modified:   src/plan/index.js
- Modified:   src/plan/plan.js
- Modified:   src/plan/plan.test.js
- Modified:   src/reporter/reporter.js
- Modified:   src/reporter/reporter.test.js